new/sidecar/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
torch>=2.4.0
faster-whisper>=1.0.3
nemo-toolkit>=1.23.0
pydantic>=2.8.0
